refactor: extract sop settings section
